Clarified some terminology
This project isn't a "port" per se, since both the source and the destination language are the same.

This project was originally a fork, but has now changed so drastically that a majority of the original code has been altered.
It would be safe to say that this project is now a project on its own and merely bases its core logic on gjtorikian's project.